@@ -1,4 +1,4 @@
-import express = require('express')
+import * as express from 'express'
 import {postToWorker} from "../modules/worker/worker-factory";
 import {removeOverrides, unHideAll} from "../modules/routes/margin";
 import {Fees} from "../modules/fees";
@@ -6,26 +6,27 @@ import {Packaging} from "../modules/packaging";
 import {Postage} from "../modules/postage";
 import ProcessMargins from "../modules/margin-calculation";
 
-const router = express.Router()
 
-router.post('/Update', async(req, res) => {
+const marginRoutes = express.Router()
+
+marginRoutes.post('/Update', async (req, res) => {
     let data = req.body?.SKU ? {skus: `'${req.body.SKU}'`} : {}
     let result = await postToWorker("update",
-        { type: "channelPrices", data: data, msg: "", reqId: "", id: new Date().getTime().toString() }
+        {type: "channelPrices", data: data, msg: "", reqId: "", id: new Date().getTime().toString()}
     )
     res.send(result.data ? result.data : {status: 'done'})
 })
-router.post('/UnHideAll', async(req, res) => {
+marginRoutes.post('/UnHideAll', async (req, res) => {
     await unHideAll()
     res.end()
 })
 
-router.post('/RemoveOverrides', async(req, res) => {
+marginRoutes.post('/RemoveOverrides', async (req, res) => {
     await removeOverrides()
     res.end()
 })
 
-router.post('/TestItem', async (req, res) => {
+marginRoutes.post('/TestItem', async (req, res) => {
     /* Requires obj in req body:
     * {
     *   postage:1
@@ -45,4 +46,5 @@ router.post('/TestItem', async (req, res) => {
     await ProcessMargins(req.body, fees, packaging, postage)
     res.send(req.body)
 })
-export = router
+
+export default marginRoutes
