@@ -1,13 +1,11 @@
-import express = require('express')
+import express from 'express'
 import {postToWorker} from "../modules/worker/worker-factory";
 
-const router = express.Router()
-
-router.post('/Online-Sales', async(req, res) => {
+const reportsRoutes = express.Router()
+reportsRoutes.post('/Online-Sales', async (req, res) => {
     let result = await postToWorker("reports",
-        { type: "online-sales", data: {}, msg: "", reqId: "", id: new Date().getTime().toString() }
+        {type: "online-sales", data: {}, msg: "", reqId: "", id: new Date().getTime().toString()}
     )
     res.send(result.data ? result.data : {status: 'done'})
 })
-
-export = router
+export default reportsRoutes
